requirements.txt
python-telegram-bot==13.15
python-dotenv==1.0.0
google-generativeai==0.1.0